+++ jsolex-core/src/main/java/me/champeau/a4j/jsolex/processing/stretching/AutohistogramStrategy.java
@@ -15,10 +15,8 @@
  */
 package me.champeau.a4j.jsolex.processing.stretching;
 
-import me.champeau.a4j.jsolex.processing.sun.BackgroundRemoval;
 import me.champeau.a4j.jsolex.processing.sun.Broadcaster;
 import me.champeau.a4j.jsolex.processing.sun.tasks.ImageAnalysis;
-import me.champeau.a4j.jsolex.processing.sun.workflow.AnalysisUtils;
 import me.champeau.a4j.jsolex.processing.util.Constants;
 import me.champeau.a4j.jsolex.processing.util.Histogram;
 import me.champeau.a4j.jsolex.processing.util.ImageWrapper32;
@@ -35,18 +33,27 @@ public final class AutohistogramStrategy implements StretchingStrategy {
     private static final double TARGET_PEAK = 0.65;
     private static final double LIMB_MARGIN = 1.1;
     private static final double REDUCED_AMPLIFICATION_LIMIT = 1.3;
-    private static final int TARGET_PEAK_REDUCTION_THRESHOLD = 8000;
 
+    public static final double DEFAULT_AMPLIFICATION_THRESHOLD = 0.95;
     public static final double DEFAULT_GAMMA = 1.5;
 
     private final double gamma;
+    private final double amplificationThreshold;
 
     private Broadcaster broadcaster;
 
     public AutohistogramStrategy(double gamma) {
+        this(gamma, DEFAULT_AMPLIFICATION_THRESHOLD);
+    }
+
+    public AutohistogramStrategy(double gamma, double amplificationThreshold) {
         if (gamma < 1) {
             throw new IllegalArgumentException("Gamma must be greater than 1");
         }
+        if (amplificationThreshold<0 || amplificationThreshold>1) {
+            throw new IllegalArgumentException("Amplification threshold must be between 0 and 1");
+        }
+        this.amplificationThreshold = amplificationThreshold;
         this.gamma = gamma;
     }
 
@@ -74,59 +81,65 @@ public final class AutohistogramStrategy implements StretchingStrategy {
         var width = image.width();
         if (ellipse.isPresent()) {
             var e = ellipse.get();
-            disk = BackgroundRemoval.neutralizeBackground(disk, 4);
             diskData = disk.data();
-            // the initial strech is to adjust the global brightness by estimating the brightness of
-            // the disk itself
-            stats = ImageAnalysis.masked(diskData, width, height, e);
+            // the initial strech is to adjust the global brightness by estimating the brightness of the disk itself
             var cx = e.center().a();
             var cy = e.center().b();
             var semiAxis = e.semiAxis();
             var radius = (semiAxis.a() + semiAxis.b()) / 2;
-            var amplificationThreshold = 0.5 * AnalysisUtils.estimateBlackPoint(disk, e);
-            if (stats.avg() > stats.max() / 8) {
-                float max = 1e-7f;
-                for (float v : diskData) {
-                    max = Math.max(v, max);
+            var amplificationThreshold = findAmplificationThreshold(diskData, width, height, e);
+            float max = 1e-7f;
+            for (float v : diskData) {
+                max = Math.max(v, max);
+            }
+            for (int y = 0; y < height; y++) {
+                for (int x = 0; x < width; x++) {
+                    var idx = y * width + x;
+                    var v = diskData[idx];
+                    float normalized = v / max;
+                    var dist = normalizedDistanceToCenter(x, y, cx, cy, radius);
+                    var gammaCorrected = (float) Math.pow(normalized, gamma) * Constants.MAX_PIXEL_VALUE;
+                    if (dist <= 1 || v < amplificationThreshold) {
+                        diskData[idx] = gammaCorrected;
+                    } else {
+                        float corrected = (float) Math.pow((gammaCorrected - amplificationThreshold) / max, Math.max(.5, gamma * (.2 * Math.pow(1 / dist, .5))));
+                        float newValue = corrected * Constants.MAX_PIXEL_VALUE;
+                        if (Float.isNaN(newValue)) {
+                            newValue = gammaCorrected;
+                        }
+                        var diff = newValue - gammaCorrected;
+                        float limitFactor = 1f;
+                        if (dist < LIMB_MARGIN) {
+                            var k = (LIMB_MARGIN - dist) / (LIMB_MARGIN - 1);
+                            limitFactor = (float) (1f - k);
+                        }
+                        if (dist > REDUCED_AMPLIFICATION_LIMIT) {
+                            limitFactor /= (float) (4 * (1 + (dist - REDUCED_AMPLIFICATION_LIMIT)));
+                        }
+                        newValue = gammaCorrected + diff * limitFactor;
+                        diskData[idx] = newValue;
+                    }
                 }
+            }
+            var newStats = ImageAnalysis.masked(diskData, width, height, e);
+            if (newStats.avg() < 4096) {
+                disk = image.copy();
+                diskData = disk.data();
+                var clamp = clamping(diskData);
+                clamp.stretch(disk);
                 for (int y = 0; y < height; y++) {
                     for (int x = 0; x < width; x++) {
                         var idx = y * width + x;
-                        var v = diskData[idx];
-                        float normalized = v / max;
-                        var dist = normalizedDistanceToCenter(x, y, cx, cy, radius);
-                        var gammaCorrected = (float) Math.pow(normalized, gamma) * Constants.MAX_PIXEL_VALUE;
-                        if (dist <= 1 || v < amplificationThreshold) {
-                            diskData[idx] = gammaCorrected;
-                        } else {
-                            float corrected = (float) Math.pow((gammaCorrected - amplificationThreshold) / max, Math.max(.5, gamma * (.2 * Math.pow(1 / dist, .5))));
-                            float newValue = (corrected * Constants.MAX_PIXEL_VALUE);
-                            if (Float.isNaN(newValue)) {
-                                newValue = gammaCorrected;
-                            }
-                            var diff = newValue - gammaCorrected;
-                            float limitFactor = 1f;
-                            if (dist < LIMB_MARGIN) {
-                                var k = (LIMB_MARGIN - dist) / (LIMB_MARGIN - 1);
-                                limitFactor = (float) (1f - k);
-                            }
-                            if (dist > REDUCED_AMPLIFICATION_LIMIT) {
-                                limitFactor /= (float) (4 * (1 + (dist - REDUCED_AMPLIFICATION_LIMIT)));
-                            }
-                            newValue = gammaCorrected + diff * limitFactor;
-                            diskData[idx] = newValue;
+                        float v = diskData[idx];
+                        if (e.isWithin(x, y) || v > 2*amplificationThreshold) {
+                            diskData[idx] = (float) Math.pow(v, .8);
                         }
                     }
                 }
-            } else {
-                disk = image.copy();
-                diskData = disk.data();
-                for (int i = 0; i < diskData.length; i++) {
-                    float v = diskData[i];
-                    diskData[i] = (float) Math.pow(v, 0.8);
-                }
-                LinearStrechingStrategy.DEFAULT.stretch(disk);
+                new ClaheStrategy(16, 256, 1.0).stretch(disk);
                 LOGGER.warn(message("skip.gamma.stretch"));
+                System.arraycopy(diskData, 0, image.data(), 0, diskData.length);
+                return;
             }
 
             System.arraycopy(diskData, 0, image.data(), 0, diskData.length);
@@ -151,6 +164,24 @@ public final class AutohistogramStrategy implements StretchingStrategy {
         for (int i = 0; i < diskData.length; i++) {
             diskData[i] = (float) (0.75 * diskData[i] + 0.25 * claheData[i]);
         }
+    }
+
+    private static ContrastAdjustmentStrategy clamping(float[] diskData) {
+        var limit = Histogram.of(diskData, 65536).cumulative().percentile(0.999);
+        return new ContrastAdjustmentStrategy(0, limit);
+    }
+
+    private double findAmplificationThreshold(float[] diskData, int width, int height, Ellipse e) {
+        var builder = Histogram.builder(65536);
+        for (int y = 0; y < height; y++) {
+            for (int x = 0; x < width; x++) {
+                if (!e.isWithin(x, y)) {
+                    builder.record(diskData[y * width + x]);
+                }
+            }
+        }
+        var h = builder.build().cumulative();
+        return h.percentile(amplificationThreshold);
     }
 
     // computes the distance to the circle center, relative to the radius. A negative value